blast-results : split out blast-results-view
 ... in preparation for adding row selection via checkboxes, etc.
Move from blast-results to blast-results-view : columnsKeyString, viewFeaturesFlag, containerStyle (factored from .hbs to a CP), dataMatrix, dataFeatures, blockNames, dataMatrixEffect, didRender(), viewFeaturesEffect, activeEffect, shownBsTab(), showTable(), createTable(), clearTable(), validateData()
