+++ frontend/app/templates/components/panel/upload/blast-results-view.hbs
@@ -1,0 +1,16 @@
+{{dataMatrixEffect}}
+{{activeEffect}}
+{{viewFeaturesEffect}}
+
+{{!-- --------------------------------------------------------------------- --}}
+
+<div id={{search.tableId}} class="actual-row"  style={{this.containerStyle}}></div>
+
+{{!-- --------------------------------------------------------------------- --}}
+
+<span class="filter-group-col">
+  {{input type="checkbox" name="viewFeaturesFlag" checked=viewFeaturesFlag }}
+  <label>Show Features (triangles)</label>
+</span>
+
+{{!-- --------------------------------------------------------------------- --}}
